program for simple calculator
A calculator is an electronic device that is used to perform arithmetic operations like addition, subtraction, multiplication, division, etc. We can create a simple calculator program in C to perform all these basic calculations. In this article, we will see how to create a simple calculator in C using the switch statement.
This program takes an arithmetic operator (+, -, *, /) and two operands as the input. It then checks whether the user wants to quit the program if yes then quit it, for this, we can use a special character and tell the user about it, like here we used “x”. The operation is performed based on the operator.

Approach
All the following steps will be performed inside a never-ending loop so that the calculator program keeps on working.
Take input of operator and then operands.
Check whether the user wants to quit the program if yes then quit it, for this, we can use a special character and tell the user about it, like here we used “x”.
Using if-else statements we will check operators and do operations accordingly.
Simple Calculator Program using switch Statement
Below is the C program to implement a simple calculator using a switch case statement: